python/requirement.txt
selenium==3.141.0
pytest==4.0.2
pytest-xdist==1.24.1
Faker==1.0.5
webdriver_manager==1.7
allure-pytest==2.6.3
Appium-Python-Client==0.43
pytest-testrail==2.3.3
delayed-assert==0.3.2